backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
asyncpg==0.30.0
alembic==1.14.1
greenlet==3.1.1

# Redis
redis[hiredis]==5.2.1

# Auth
pyjwt[crypto]==2.10.1
bcrypt==4.2.1
authlib==1.4.0
httpx==0.28.1

# Scraping
beautifulsoup4==4.12.3
lxml==5.3.0

# Scheduling
apscheduler==3.10.4

# Data processing
pydantic[email]==2.10.4
pydantic-settings==2.7.1

# Email
aiosmtplib==3.0.2

# WebSocket
websockets==14.1

# Utilities
python-dotenv==1.0.1
orjson==3.10.13

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
